requirements.txt
absl-py==1.0.0
albumentations==1.1.0
argon2-cffi==21.1.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blessings==1.7
Bottleneck==1.3.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cycler==0.11.0
defusedxml==0.7.1
entrypoints==0.3
executing==0.8.2
fastai==1.0.61
fastprogress==1.0.0
filelock==3.4.0
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
gdown==4.2.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.13.3
imgaug==0.4.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
imutils==0.5.4
iniconfig==1.1.1
ipykernel==6.6.0
ipython==7.30.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
littleutils==0.2.2
Markdown==3.3.6
MarkupSafe==2.0.1
mat4py==0.5.0
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numexpr==2.8.0
numpy==1.21.4
nvidia-ml-py3==7.352.0
nvidia-smi==0.1.3
oauthlib==3.1.1
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
Pillow==8.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.19.1
psutil==5.8.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.5
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.2
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools-scm==6.3.2
Shapely==1.8.0
sklearn==0.0
sorcery==0.2.2
soupsieve==2.3.1
tb-nightly==2.8.0a20211209
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thop==0.0.31.post2005241907
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torch-tb-profiler==0.3.1
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.6.0
